move MetricDomainTypes to a common location
