+++ great_expectations/core/metric_domain_types.py
@@ -1,0 +1,11 @@
+import enum
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+class MetricDomainTypes(enum.Enum):
+    TABLE = "table"
+    COLUMN = "column"
+    COLUMN_PAIR = "column_pair"
+    MULTICOLUMN = "multicolumn"
