+++ great_expectations/expectations/metrics/column_aggregate_metrics/column_histogram.py
@@ -5,6 +5,7 @@ from typing import Any, Dict
 import numpy as np
 import pandas as pd
 
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.core.util import (
     convert_to_json_serializable,
     get_sql_dialect_floating_point_infinity_value,
@@ -14,7 +15,6 @@ from great_expectations.execution_engine import (
     SparkDFExecutionEngine,
     SqlAlchemyExecutionEngine,
 )
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.expectations.metrics.column_aggregate_metric_provider import (
     ColumnAggregateMetricProvider,
 )
