+++ great_expectations/rule_based_profiler/domain_builder/table_domain_builder.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.rule_based_profiler.domain_builder import DomainBuilder
 from great_expectations.rule_based_profiler.types import Domain, ParameterContainer
