+++ great_expectations/rule_based_profiler/types/data_assistant_result/data_assistant_result.py
@@ -13,6 +13,7 @@ from IPython.display import HTML, display
 
 from great_expectations import __version__ as ge_version
 from great_expectations.core import ExpectationConfiguration, ExpectationSuite
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.core.usage_statistics.events import UsageStatsEvents
 from great_expectations.core.usage_statistics.usage_statistics import (
     UsageStatisticsHandler,
@@ -24,7 +25,6 @@ from great_expectations.core.util import (
     in_jupyter_notebook,
     nested_update,
 )
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.rule_based_profiler.config import RuleConfig
 from great_expectations.rule_based_profiler.helpers.util import (
     get_or_create_expectation_suite,
