+++ great_expectations/rule_based_profiler/types/domain.py
@@ -4,8 +4,8 @@ from enum import Enum
 from typing import Any, Dict, Optional, Union
 
 from great_expectations.core import IDDict
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.core.util import convert_to_json_serializable
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.types import SerializableDictDot, SerializableDotDict
 from great_expectations.util import (
     deep_filter_properties_iterable,
