+++ tests/execution_engine/test_pandas_execution_engine.py
@@ -13,7 +13,7 @@ except:
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.batch_spec import RuntimeDataBatchSpec, S3BatchSpec
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.execution_engine.pandas_execution_engine import (
     PandasExecutionEngine,
     storage,
