+++ tests/execution_engine/test_sparkdf_execution_engine.py
@@ -7,8 +7,8 @@ import pytest
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.batch_spec import PathBatchSpec, RuntimeDataBatchSpec
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.execution_engine import SparkDFExecutionEngine
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.expectations.row_conditions import (
     RowCondition,
     RowConditionParserType,
