+++ tests/rule_based_profiler/conftest.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 from ruamel.yaml import YAML
 
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.rule_based_profiler import RuleBasedProfiler
 from great_expectations.rule_based_profiler.config.base import RuleBasedProfilerConfig
 from great_expectations.rule_based_profiler.domain_builder import ColumnDomainBuilder
