+++ tests/rule_based_profiler/domain_builder/test_categorical_column_domain_builder.py
@@ -4,8 +4,8 @@ import pytest
 
 from great_expectations import DataContext
 from great_expectations.core.batch import BatchRequest
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.exceptions import ProfilerConfigurationError
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.rule_based_profiler.domain_builder import DomainBuilder
 from great_expectations.rule_based_profiler.domain_builder.categorical_column_domain_builder import (
     CategoricalColumnDomainBuilder,
