+++ tests/rule_based_profiler/domain_builder/test_map_metric_column_domain_builder.py
@@ -2,7 +2,7 @@ from typing import List
 
 from great_expectations import DataContext
 from great_expectations.core.batch import BatchRequest
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.rule_based_profiler.domain_builder import (
     MapMetricColumnDomainBuilder,
 )
