+++ tests/rule_based_profiler/parameter_builder/test_parameter_container.py
@@ -1,6 +1,6 @@
 from typing import Dict, List
 
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.rule_based_profiler.types import (
     Domain,
     ParameterContainer,
