+++ tests/rule_based_profiler/parameter_builder/test_partition_parameter_builder.py
@@ -1,7 +1,7 @@
 from typing import Dict, Optional
 
 from great_expectations import DataContext
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.rule_based_profiler.helpers.util import (
     get_parameter_value_and_validate_return_type,
 )
