+++ tests/rule_based_profiler/parameter_builder/test_simple_date_format_string_parameter_builder.py
@@ -3,8 +3,8 @@ from typing import Dict
 import pytest
 
 import great_expectations.exceptions.exceptions as ge_exceptions
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.data_context import DataContext
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.rule_based_profiler.helpers.util import (
     get_parameter_value_and_validate_return_type,
 )
