+++ tests/test_fixtures/rule_based_profiler/plugins/my_custom_semantic_type_column_domain_builder.py
@@ -1,7 +1,7 @@
 from typing import List, Optional, Union
 
 from great_expectations import DataContext
-from great_expectations.execution_engine.execution_engine import MetricDomainTypes
+from great_expectations.core.metric_domain_types import MetricDomainTypes
 from great_expectations.rule_based_profiler.domain_builder import DomainBuilder
 from great_expectations.rule_based_profiler.helpers.util import (
     build_domains_from_column_names,
